# Conflicts:
#	Features/FiatConnect/Sources/Scenes/FiatScene.swift
#	Features/FiatConnect/Sources/ViewModels/FiatSceneViewModel.swift

+++ Packages/GemstonePrimitives/Sources/Extensions/GemUtxo+GemstonePrimitives.swift
@@ -19,7 +19,7 @@ extension UTXO {
     public func map() -> GemUtxo {
         GemUtxo(
             transactionId: transaction_id,
-            vout: UInt32(vout),
+            vout: vout,
             value: value,
             address: address
         )
